backend/requirements.txt
playwright>=1.58.0
pyyaml>=6.0
requests>=2.32.0
